Added Object Pickup and It doesn't work
Tried implementing a new item pickup and throw system, and it doesn't work. The whole object i pick up constantly changes its transforms and becomes a thin plank instead of the cube it was. Gonna go back to the old item throwing scripting approach... this project is dead in the water.